+++ StartUp.py
@@ -10,6 +10,7 @@ from GameClasses import *
 XRANGE = 7 #when changing these also change the values in the CreativeMode.load() function
 YRANGE = 9
 ZRANGE = 4
+DRANGE = 2
 
 def Reset():
     global MAPS1
@@ -45,6 +46,19 @@ def Reset():
     Map("Inside University Hall",(6,1,1),"~UNIVERSITY HALL~:\nThe exit is to your left.","The glares from portraits of old white founding fathers intimidate you.\nThe memories of failing midterms in their presence sends you into an almost trance-like state.\nYou notice that the portrait of Keyes totally looks like Stephen Fry.\nAfter talking yourself out of stealing a piece of Mac history you plan your next move.",('r','b','f'),True),
     Map("The Phoenix",(5,4,1),"~THE PHOENIX~:\nGeneral Sciences Building is to your left. Hamilton Hall is to your rear.\nThe trail to Bates is in front of you. You can also go down the stairs to Bridges Cafe.","Upon entering, a rush of memories from last night enter your mind.\nPeople's faces are a blur but you somewhat recall '16 tequila shots' as something you said.\nYou see a mysterious Hooded Man as he beckons you to come over.",('r'),True),
     Map("BSB",(3,4,1),"~BSB~:\nEnter JHE to your left. General Sciences building to your right.\nFront of BSB is behind you. The Police station is in front of you.\nThe Quantum Tunnel is beneath you.","You were told JHE would be your home but after picking Eng Phys\nyou didn't realize how wrong you were.\nAt least the cafe is better than JHE's.\nA shudder runs through your body as you draw nearer to the electronics labs.\nWiping a cold sweat from your brow you plan ahead.",(),True,(6,6,5)),
+    #TODO CAPSTONE ROOM, To be Moved to interrior
+    Map("Capstone Doorway",(3,4,3),"Capstone Doorway~:\n","You walk into the room that is a mess. This is a disaster,\nhow did this happen? Do these people live here?",('u'),True,1),
+    Map("Circuit Smart",(4,4,3),"Circuit Smart~:\n","(Print text of spagetti)\nIf you unplugged one wire of this these people would go insane.",('d','u'),True,1),
+    Map("Milli",(4,3,3),"Milli~:\n","Great Job Mili!",('d','u'),True,1),
+    Map("NANOrims",(4,2,3),"NANOrims~:\n","NANNNOOORYMMSS.",('d','u'),True,1),
+    Map("S.T.A.R.S.",(3,2,3),"S.T.A.R.S.~:\n","S.T.A.R.S. PLEASE WORK.\nYou see a man wearing a pink shirt and a giant robot point a cannon where ever he goes.\nIs this how the world ends?",('d','u'),True,1), #S.T.A.R.S. System T69
+    Map("FRAS",(3,3,3),"FRAS~:\n","Yeah that's a tank. This is a 3D printed Tank.",('d','u'),True,1),
+    Map("T.A. Area",(2,4,3),"T.A. Area~:\n","Is this where the PSRs get lost? Also just storage space for STARS",('d','u'),True,1),
+    Map("ZebraShark",(2,3,3),"ZebraShark~:\n","Where are they? O.m.g. is that a pool downstairs?",('d','u'),True,1),
+    Map("THE ECLIPSE",(2,2,3),"THE ECLIPSE~:\n","AW DANG The Eclipse! But yeah this windshield is a bit much",('d','u'),True,1),
+    Map("Electronics Lab",(1,4,3),"Peter's Lab~:\n","I'm just glad to not have to be in here anymore.",('d','u'),True,1),
+    Map("Peter Johnason's Office",(1,3,3),"Peter Johnason's Office~:\n","The grand sorcerer's mystic place",('d','u'),True,1),
+    
     Map("Front BSB",(3,3,1),"~FRONT OF BSB~:\nBSB field is at your rear. A road is to your right.\nThe entrance to JHE is to your left. BSB is infront of you","You look up and see the McMaster coat of arms engraved into the side of BSB.\nThis little pathway has been well worn and you wonder why they don't connect JHE and BSB anyway.\nAn underground (quantum) tunnel would save some hardship on a rainy day...",(),False),
     Map("Police Station",(3,5,1),"~POLICE STATION~:\nThe path between JHE & BSB behind you. The GO station lies ahead of you.\nThe Nuclear Research Building is to your left. The Tandem Accelerator is to your right.","Thoughts of getting kicked out of res parties fill your head.\nThose special constables are punks.\nYou mutter the lyrics of a certain N.W.A hit. After avoiding a campus PD cruiser\nscreaming around the blind corner, you think of what to do next.",(),True),
     Map("Keyes",(4,6,1),"~MARY KEYES~:\nYou can head to the Bridge if you go forward. The GO station is to your left.\nThe Tandem Accelerator is at your rear. Bates is to your right.","A snack station that's open 'til midnight.\nWhat a life saver indeed.\nYou quickly thank the engineering gods for Mary Keyes and fight yourself\nfrom ordering chicken fingers & fries super combo...",(),True),
@@ -107,19 +121,7 @@ def Reset():
     Map("Climbing MDCL",(1,0,2),"~CLIMBING MDCL~:\nYou can continue climbing or go back down...","You struggle and find yourself on the outer wall of MDCL.",('f','b','l','r'),True),
     Map("Roof of MDCL",(1,0,3),"~MDCL Roof~:\nYou can only go back down.","With all of your might you pull yourself up and onto the roof.\nYou can see the majority of campus from up here!",('f','b','l','r','u'),False),
     Map("3rd Floor Thode",(1,7,3),"~3RD FLOOR THODE~:\nYou can only go back down the stairs.","After hauling up another set of stairs you realize you probably shouldn't\nhave given up on your Pulse membership...\nThe sounds of sobbing can be heard from all around you.\nAfter you dispense all of your tissues to passing I-Sci's you plan your next move.",('f','b','l','r','u'),True),
-    Map("2nd Floor Thode",(1,7,2),"~2ND FLOOR THODE~:\nYou can go up or down the stairs.","As you enter Club Thode the smell of feet enters your nostrils.\nWho goes barefoot in a library? C'mon.\nThe memory of countless hours spent slamming together a report made of nonsense, caffeine,\nand hope makes you light headed...\nComing to your senses... You plan your next move.",('f','b','l','r'),True),
-    #TODO CAPSTONE ROOM, To be Moved to interrior
-    Map("Capstone Doorway",(3,4,3),"Capstone Doorway~:\n","You walk into the room that is a mess. This is a disaster,\nhow did this happen? Do these people live here?",('u'),True,"bsb"),
-    Map("Circuit Smart",(4,4,3),"Circuit Smart~:\n","(Print text of spagetti)\nIf you unplugged one wire of this these people would go insane.",('d','u'),True,"bsb"),
-    Map("Milli",(4,3,3),"Milli~:\n","Great Job Mili!",('d','u'),True,"bsb"),
-    Map("NANOrims",(4,2,3),"NANOrims~:\n","NANNNOOORYMMSS.",('d','u'),True,"bsb"),
-    Map("S.T.A.R.S.",(3,2,3),"S.T.A.R.S.~:\n","S.T.A.R.S. PLEASE WORK.\nYou see a man wearing a pink shirt and a giant robot point a cannon where ever he goes.\nIs this how the world ends?",('d','u'),True,"bsb"), #S.T.A.R.S. System T69
-    Map("FRAS",(3,3,3),"FRAS~:\n","Yeah that's a tank. This is a 3D printed Tank.",('d','u'),True,"bsb"),
-    Map("T.A. Area",(2,4,3),"T.A. Area~:\n","Is this where the PSRs get lost? Also just storage space for STARS",('d','u'),True,"bsb"),
-    Map("ZebraShark",(2,3,3),"ZebraShark~:\n","Where are they? O.m.g. is that a pool downstairs?",('d','u'),True,"bsb"),
-    Map("THE ECLIPSE",(2,2,3),"THE ECLIPSE~:\n","AW DANG The Eclipse! But yeah this windshield is a bit much",('d','u'),True,"bsb"),
-    Map("Electronics Lab",(1,4,3),"Peter's Lab~:\n","I'm just glad to not have to be in here anymore.",('d','u'),True,"bsb"),
-    Map("Peter Johnason's Office",(1,3,3),"Peter Johnason's Office~:\n","The grand sorcerer's mystic place",('d','u'),True,"bsb")]
+    Map("2nd Floor Thode",(1,7,2),"~2ND FLOOR THODE~:\nYou can go up or down the stairs.","As you enter Club Thode the smell of feet enters your nostrils.\nWho goes barefoot in a library? C'mon.\nThe memory of countless hours spent slamming together a report made of nonsense, caffeine,\nand hope makes you light headed...\nComing to your senses... You plan your next move.",('f','b','l','r'),True)]
 
     #Items: Equipment.name = "Name" - Equipment.location = tuple of location - Equipment.image = .jpg of item
     #       Equipment.info = "info" - Equipment.worn = 'head','hand','body',or 'off-hand' - Equipment.stats = (Atk,Def,Spd)
